Refactor Parsec Git into GitParse module
Its funny how you get stuck on things when learning new PLs, and,
a lot of the time, you do things just to get around the fact that
you can't figure them out.

Here, I am unable to figure out how to work with the <|> operator
from both Turtle and Parsec. Ideally, I should be able to name one, or
rename one, or something, but I can't figure it out.

It needs to be in its own module anyway, but I think this should
probably be in a Git/ParseOutput module, but i don't really know how
to do all that yet, so to keep it simple for now its just in a module
called "GitParse"